[WIP] Adding updated Field/FieldView based on contenteditable
Introducing a new Field/FieldView combination whose view is based on
the HTML5 contenteditable property. This allows us to have much more
fine grained control over the component, in particular how text lines
are handled.

For the moment, there is basic and broken functionality as of this
commit. Note that preloading the text of the scriptEditor is not
currently working.

TODO: Figure out how to wrap the different lines in the FieldView with
html tags. This allows us to perform the selection behaviors when the
autoSelect and lockText properties are both true, as in normal
HyperCard, thus giving us a selection list.