Updating ohm tests so they run when we npm test
Previously, any of the tests in js/ohm/tests were not running as a
part of the normal overall testing script. We update that script to
include these tests.

I have also renamed the two ohm test files so that they better conform
to the pattern we use in other JS tests.