Creating 'descriptor' objects for delayed evaluation]
-- Prelude
We have a need for -- and have already partially implemented --
mechanisms for delayed evaluation. For example, we might need to pass
Part instances to command handlers as arguments, but we do not want to
send those objects along in the messages since they are not fully
serializable/deserializable. Similarly, we have variables that should
only be looked up at command function runtime, but we'd need some
extra information about what variable name to lookup.

This is where 'descriptors' come in. These are simply JSONable objects
that have some extra information needed that the compiler's
`evaluate()` function can case through and take appropriate action
upon.

I have created the new file `descriptors.js` to store these initial
descriptor objects. For the moment, we are only making use of the
STVariable descriptor.

- The most important thing to note is that we need to use
Object.assign() when creating new copies of these descriptor objects.