[WIP] Updating grammar, semantics, and tests for variable / setting
-- About
This commit is a WIP. It deals with adjustments to the grammar and
semantics for not only reading and writing local variables, but also
makes some important adjustments to the setProperty grammar,
semantics, and (eventually) how the command itself is handled.

To start, we now divide up the Command_setProperty in the grammar to
more constituent parts. These will help the parsing and also help
return more concrete objects from semantic parsing. I have written
additional tests for this decomposition and they are all passing.

Next, I've added another test module that will attempt to do the
following integration, high level:
1) Test that we can "put <anyLiteral> into <variableName>" correctly
and without error (PASSING);
2) That we can set a variable as in step 1, in a script, and then also
set the property of some Part to whatever the variable is storing
during execution. The plan is to use a button, its click handler, and
to set the button's name property (NOT PASSING)

In attempting to get the above Step 2 to pass, I've realized that the
setProperty command handling -- both at the System level and in the
constituent Parts -- will need to be heavily refactored. Part of the
current roundabout setup was dealing with preserving "context", but
between the new message `senders` array (and originalSender) and the
new decomposed grammar for set property targets, we should not need
the extra complexity.

TODO: Refactor setProperty handling in System and Part levels.