[WIP] Refactoring newModel command grammar/semantics/implementation
This is similar to work that was done with the setProperty grammar,
semantics, and implementations. Tests are still failing because many
of them are quite specific, and it's still unclear what should be
allowed or not (or what things "mean") in certain scripted add
commands.

I also recommend that we separate the newModel command message from
the system's newModel method. They are used in two different ways:
internally (not as a message send) and externally (via a message
send). These are two different contexts, and that is important: the
former will not have any information about what Part sent the message
and may not know where to add the new part.

Additionally, I think newModel is inappropriate as a scriptable
command, and that we should have a separate command called addPart or
similar which down the line resolves to a newModel call as it exists
in System, but with all the context, target, and additional
information appropriately configured.