Initial commit of VisionField Part/View
-- Description
This is a new experimental part for integrating with the Vision tool
that Thomas is currently developing.

We introduce a very basic Part/View combo that has the ability to poll
a vision url and to draw any circles from coordinates as they come in.

Key properties to set on the model are:
* polling -- when set to true, will ask the vision server at the src
property url for new coordinates
* coordinates -- an array of coordinates supplied by the vision server
or set manually
* src -- the url of the vision server
* pollTime -- the delay, in milliseconds, between polling
intervals (ie how often the vision server should be polled)