VisionWorker and WorldStack cursor integration
-- About
This commit deals with an initial version of a vision-based cursor on
the WorldStack.

In order to make this work, we have created a WebWorker called
VisionWorker which, when initialized, will spin up a separate thread
that can handle making vision server requests and passing messages
back to the WorldStack.

WorldStack, when initialized, will create and instance of this worker
but it is not turned on to 'listening mode' by default. To do so, the
user must set the WorldStack's `useVisionCursor` partProperty to true,
which will send the appropriate start message to the VisionWorker.

All communications with the VisionWorker are message based and should
be completely handled within the WorldStack part's methods. The
WorldView and any other concerned objects in the ST system should
simply use PartProperties for updating themselves. For example, the
WorldView reacts to `useVisionCursor` and `cursorCoordinate`
partProperties of its model, which is enough information to toggle the
cursor display and update where it should appear on the screen.

-- Live Example
To get this running, simply open the bootstrap file and add this code
to the browser console:

```
let worldView = document.querySelector('st-world');
let world = worldView.model;
world.partProperties.setPropertyNamed(world, 'useVisionCursor', true);
```

If you have a Vision server running on localhost:5000, and circle
drawing is enabled and working, you should see a red square cursor
moving on top of the WorldStack (ie the viewport)